+++ client/src/componets/authentication/forgetPassword/ForgetPassword.jsx
@@ -1,16 +1,12 @@
 import axios from "axios";
 import { useState } from "react";
-import { useNavigate,Link,Outlet } from "react-router-dom";
+import { useNavigate} from "react-router-dom";
 import { emailValid } from '../../../utilities/validetion/validetion.js';
-
-import CloseIcon from '@mui/icons-material/Close';
-import {Box, Typography, TextField, Button }from '@mui/material';
-// import '../auth/Auth.css';
+import {Box, TextField, Button }from '@mui/material';
 
 export default function ForgetPassword({handleResetPass}) {
 
 
-const navigate = useNavigate();
 const [email,setEmail]= useState('');
 
 const submit =async()=>{
@@ -21,7 +17,6 @@ const submit =async()=>{
       try {
         const answer = await axios.post('http://localhost:3005/auth/forget-pass', { email })
         handleResetPass()
-        // navigate('/reset-pass');
 
     }
     catch (err) {
@@ -49,9 +44,6 @@ const submit =async()=>{
             }}
             variant="outlined"
           >
-            <div>
-            
-            </div>
             <TextField
               required
             id='email'
