+++ client/src/componets/home/Home.jsx
@@ -9,50 +9,23 @@ import UserConnected from '../../context/UserConnected';
 import {Box} from '@mui/material'
 
 
-export default function Home() {
-
-    
-    const { userConnected, setUserConnected } = useContext(UserConnected);
-    const [anchorUserMenu, setAnchorUserMenu] = useState(null);
+export default function Home() {   
     const [openLogIn, setOpenLogIn] = useState(false);
     const [authPopup, setAuthPopup] = useState(false);
-    
-
-
     const handleAuthPopup = () => {
         setAuthPopup((authPopup) => !authPopup)
     }
-    // const handleOpenLogIn = () => {
-    //     setOpenLogIn(true);
-    //     handleCloseUserMenu();
 
-    // };
     const handleCloseLogIn = () => {
         setOpenLogIn(false);
     };
-    // const handleAuth = () => {
-    //     setUserConnected(!userConnected)
-    // }
-    // const handleOpenUserMenu = (event) => {
-    //     setAnchorUserMenu(event.currentTarget);
-    // };
-
-    // const handleCloseUserMenu = () => {
-    //     setAnchorUserMenu(null);
-    // };
-
-    // const handleLogOut = () => {
-    //     setUserConnected(null);
-    // }
-    // const aaa = () => {
-    //     setUserConnected({ name: 'amit' })
-    // }
     
     return (
         <Box sx={{ flexGrow: 1}}>
             <AuthPopupContex.Provider value={{ handleAuthPopup }}>
                 <BrowserRouter>
                 <Nav setOpenLogIn={setOpenLogIn}/>
+                
                 <Routes>
                     <Route path='/' element={<BodyApp openLogIn={openLogIn} handleCloseLogIn={handleCloseLogIn} />} />
                     <Route path='/about' element={<About/>} />
