+++ client/src/componets/home/bodyApp/BodyApp.jsx
@@ -1,14 +1,16 @@
-
 import React, { useContext, useEffect } from "react";
 import axios from "axios";
 import UserConnected from "../../../context/UserConnected";
 import Auth from "../../authentication/auth/Auth";
-import Post from "../../post/Post";
+import CreatePost from "../CreatePost/CreatePost";
+import PostButton from "./PostButton";
+import Card from "@mui/material/Card";
+import CardActions from "@mui/material/CardActions";
+import CardContent from "@mui/material/CardContent";
+import CardMedia from "@mui/material/CardMedia";
 
-import Card from '@mui/material/Card';
-import CardActions from '@mui/material/CardActions';
-import CardContent from '@mui/material/CardContent';
-import CardMedia from '@mui/material/CardMedia';
+
+import Posts from "./Posts/Posts";
 
 import {
   DialogActions,
@@ -31,48 +33,86 @@ import { useState } from "react";
 import { typography } from "@mui/system";
 const options = ["English", "Math", "Program engineer"];
 
-
 export default function BodyApp({ openLogIn, handleCloseLogIn }) {
-
   const [option, setOption] = useState([]); ////////////////////
 
-  const [posts, setPosts] = useState(
-    [
-      { auther_name: "david", category: "Languages", id: 1, post: "a", sub_category: "English", time_from: "2001-01-31T22:00:00.000Z", time_to: "2001-02-28T22:00:00.000Z", user_id: 1 },
-      { auther_name: "david", category: "Languages", id: 2, post: "a", sub_category: "English", time_from: "2001-01-31T22:00:00.000Z", time_to: "2001-02-28T22:00:00.000Z", user_id: 1 },
-      { auther_name: "david", category: "Languages", id: 3, post: "a", sub_category: "English", time_from: "2001-01-31T22:00:00.000Z", time_to: "2001-02-28T22:00:00.000Z", user_id: 1 },
-      { auther_name: "david", category: "Languages", id: 4, post: "a", sub_category: "English", time_from: "2001-01-31T22:00:00.000Z", time_to: "2001-02-28T22:00:00.000Z", user_id: 1 }
-    ]
-  );
+  const [posts, setPosts] = useState([
+    {
+      auther_name: "david",
+      category: "Languages",
+      id: 2,
+      post: "a",
+      sub_category: "English",
+      date: "01/01/2023",
+      time_from: "00:00",
+      time_to: "00:00",
+      user_id: 1,
+    },
+    {
+      auther_name: "david",
+      category: "Languages",
+      id: 3,
+      post: "a",
+      sub_category: "English",
+      date: "01/01/2023",
+      time_from: "00:00",
+      time_to: "00:00",
+      user_id: 1,
+    },
+    {
+      auther_name: "david",
+      category: "Languages",
+      id: 1,
+      post: "a",
+      sub_category: "English",
+      date: "01/01/2023",
+      time_from: "00:00",
+      time_to: "00:00",
+      user_id: 1,
+    },
+    {
+      auther_name: "david",
+      category: "Languages",
+      id: 4,
+      post: "a",
+      sub_category: "English",
+      date: "01/01/2023",
+      time_from: "00:00",
+      time_to: "00:00",
+      user_id: 1,
+    },
+  ]);
 
   const { userConnected, setUserConnected } = useContext(UserConnected);
   const [openPost, setOpenPost] = useState(false);
   const [value, setValue] = useState(null);
   const [inputValue, setInputValue] = useState("");
-  const aaa = () => {   ////test
-    setUserConnected({ 
-      name:'Moshe',
-      email: 'Moshe@gmail.com',
-      country: 'Israel',
-      languages:'Hebrew',
-      age: '30',
-      phone_number: '052434343', });
+  const aaa = () => {
+    ////test
+    setUserConnected({
+      name: "Moshe",
+      email: "Moshe@gmail.com",
+      country: "Israel",
+      languages: "Hebrew",
+      age: "30",
+      phone_number: "052434343",
+    });
   };
 
   useEffect(() => {
     getPosts();
-
   }, []);
 
   const getPosts = async () => {
     try {
-      const postsList = await (await axios.get('http://localhost:3005/post/get-all')).data
-      setPosts(postsList)
-    }
-    catch (err) {
+      const postsList = await (
+        await axios.get("http://localhost:3005/post/get-all")
+      ).data;
+      setPosts(postsList);
+    } catch (err) {
       console.log(err);
     }
-  }
+  };
   const BootstrapDialog = styled(Dialog)(({ theme }) => ({
     "& .MuiDialogContent-root": {
       padding: theme.spacing(2),
@@ -109,10 +149,9 @@ export default function BodyApp({ openLogIn, handleCloseLogIn }) {
     onClose: PropTypes.func.isRequired,
   };
 
-
   return (
     <Box>
-      <></>
+
       <BootstrapDialog
         onClose={handleCloseLogIn}
         aria-labelledby="customized-dialog-title"
@@ -132,60 +171,30 @@ export default function BodyApp({ openLogIn, handleCloseLogIn }) {
         <Typography variant="h3">user not Connected</Typography>
       )}
       <Button onClick={aaa}>test connected</Button>
-      <Box sx={{ display: 'flex', flexWrap: 'wrap' }}>
-        {posts ? posts.map((post, index) => {
-
-          return (
-
-            <Card key={index} sx={{
-              margin: '20px',
-              maxWidth: 400,
-            }}>
-              <CardMedia
-                component="img"
-                alt={post.category}
-                height="100"
-                image="https://images.unsplash.com/photo-1512917774080-9991f1c4c750?auto=format&w=350&dpr=2"
-                // src="./joachim-schnurle-OOEKfjCRBWU-unsplash.jpg"
-              />
-              <CardContent>
-                <Typography gutterBottom variant="h5" component="div">
-                  {post.auther_name}
-
-                </Typography>
-                <Typography variant="body1" color="text.secondary">category: {post.category} </Typography>
-                <Typography variant="body1" color="text.secondary">sub category: {post.sub_category} </Typography>
-                <Typography variant="body1" color="text.secondary">i am able in time: </Typography>
-                <Typography variant="body3">{post.time_from} </Typography>
-                <Typography variant="body3">{post.time_to} </Typography>
-
-              </CardContent>
-              <CardActions>
-                <Button size="small">Ask To</Button>
-                <Button size="small">Learn More</Button>
-              </CardActions>
-            </Card>
-
-
-          )
-
-        }) : null}
-
+      <Box>
+      <Posts/>
       </Box>
       {userConnected ? (
         <Fab
-          sx={{ position: "absolute", bottom: 50, right: 50 }}
+          sx={{ position: "fixed", bottom: 20, left: 30 }}
           variant="extended"
           color="primary"
           onClick={() => {
             setOpenPost(true);
           }}
         >
-          <AddIcon />POST
+          <AddIcon />
         </Fab>
+      ) : // <PostButton />
+      null}
+      {openPost ? (
+        <CreatePost
+          open={openPost}
+          setOpen={setOpenPost}
+          option={option}
+          setOption={setOption}
+        />
       ) : null}
-      {openPost ? <Post open={openPost} setOpen={setOpenPost} option={option} setOption={setOption}/> : null}
-
     </Box>
   );
 }
