+++ client/src/componets/home/bodyApp/PostButton.jsx
@@ -1,0 +1,14 @@
+import Fab from '@mui/material/Fab'
+import AddIcon from "@mui/icons-material/Add";
+export default function BodyApp(setOpenPost){<Fab
+          sx={{ position: 'fixed', 
+          bottom: 20,
+          left: 30, }}
+          variant="extended"
+          color="primary"
+          onClick={() => {
+            setOpenPost(true);
+          }}
+        >
+          <AddIcon />
+        </Fab>}
